makes QuestionSuggestions its own component, sets it as forwardRef
