+++ components/questions/QuestionSuggestions.tsx
@@ -1,0 +1,75 @@
+import { ForwardedRef, forwardRef, useState } from "react"
+
+interface QuestionSuggestionsProps {
+  chooseSuggestion: (suggestion: string) => void
+}
+
+// eslint-disable-next-line @typescript-eslint/naming-convention
+const QuestionSuggestions = forwardRef<
+  HTMLDivElement,
+  QuestionSuggestionsProps
+>(({ chooseSuggestion }, ref: ForwardedRef<HTMLDivElement>) => {
+  const suggestions = [
+    "Will GPT-5 be released before Jan 2025?",
+    "Will I write a blog post this week?",
+    "Will volunteering abroad make me all-things-considered happier?",
+    "Each day I’ll write down whether I want to leave or stay in my job. After 2 months, will I have chosen ‘leave’ on >30 days?",
+    "Will I judge that AI was a major topic of debate in the US election?",
+    "Will I finish my todo list today?",
+    "Will our user satisfaction rating exceed 8.0/10?",
+    "Will I win my next game of Agricola?",
+    "Will Our World in Data report that >5% of global deaths are due to air pollution by 2030?",
+    "Will I still be discussing my fear of flying with my therapist in 2025?",
+    "If we choose this HR provider, will I think it was a good idea in two month’s time?",
+    "Will AMF be funding-constrained this year?",
+    "Will I have a child by Jan 2025?",
+    "Will my mentor agree that pivoting now was the right choice?",
+    "Will I meditate every day this week?",
+    "Will the rest of the team prefer this redesign to the current layout?",
+    "Will anyone on the animal advocacy forum share evidence that convinces me that abolitionist protests are net-beneficial?",
+    "Will >80% of my Twitter followers agree that I should keep the beard?",
+    "On December 1st, will Marco, Dawn, and Tina all agree that the biosecurity bill passed without amendments that removed its teeth?",
+    "If I survey 40 random Americans online, will our current favourite name be the most popular?",
+  ]
+
+  const [showAll, setShowAll] = useState(false)
+
+  return (
+    <div
+      ref={ref}
+      className="w-full bg-white shadow-inner rounded-b-md px-6 pt-4 pb-6 mb-6 flex flex-col items-start gap-2 z-10"
+    >
+      <h4 className="select-none pl-4">{"Here are a few ideas..."}</h4>
+      {suggestions.slice(0, showAll ? undefined : 8).map((suggestion) => (
+        <button
+          key={suggestion}
+          className="btn btn-ghost text-left text-neutral-500 font-normal leading-normal"
+          onClick={(e) => {
+            chooseSuggestion(suggestion)
+            e.preventDefault()
+          }}
+        >
+          <span className="ml-4">
+            <span className="text-neutral-500 font-semibold mr-2 -ml-4">+</span>
+            <span>{suggestion}</span>
+          </span>
+        </button>
+      ))}
+      {!showAll && (
+        <button
+          className="btn"
+          onClick={(e) => {
+            e.preventDefault()
+            setShowAll(true)
+          }}
+        >
+          Show more
+        </button>
+      )}
+    </div>
+  )
+})
+
+QuestionSuggestions.displayName = "QuestionSuggestions"
+
+export default QuestionSuggestions
